requirements.txt
Babel==1.3
argparse==1.2.1
cffi==0.8.6
cliff==1.8.0
cmd2==0.6.7
cryptography==0.6.1
iso8601==0.1.10
jsonpatch==1.9
jsonpointer==1.5
jsonschema==2.4.0
netaddr==0.7.12
oslo.config==1.4.0
oslo.i18n==1.0.0
oslo.serialization==1.0.0
oslo.utils==1.0.0
pbr==0.10.0
prettytable==0.7.2
pyOpenSSL==0.14
pycparser==2.10
pyparsing==2.0.3
python-glanceclient==0.14.2
python-keystoneclient==0.11.2
python-neutronclient==2.3.9
python-novaclient==2.20.0
pytz==2014.9
requests==2.4.3
simplejson==3.6.5
six==1.8.0
stevedore==1.1.0
warlock==1.1.0
wsgiref==0.1.2
